// SutraPlusUI-main/src/app/admin/components/sales/models/dc-summary.model.ts
export class DCSummary {
  BNo: number;
  CompanyId: number;
  Date: Date;
  LorryNumber: string;
  SenderDetails: string;
  ReceiverDetails: string;
  DeliveryDetails: string;
  DeliveryAddress: string;
  ConsignorsName: string;
  ConsignorsAddress: string;
  ConsigneesName: string;
  ConsigneesAddress: string;
  FromLocation: string;
  TotalFrieght: string;
  Advance: string;
  Balance: string;
  ToLocation: string;
}